ci: split CustomerService file, add ConfigureAwait, workaround IL2091
Three CI failures addressed:
- MA0048: move CustomerService into its own file
- MA0004: add ConfigureAwait(false) on every await
- IL2091: NoWarn with tracking reference to #7. The generator's
  emitted Add{Service}Cache<TImpl> extension is missing
  [DynamicallyAccessedMembers(PublicConstructors)] on TImpl,
  propagating an ILC warning from AddTransient<T>. Real bug,
  separately fixed.
