+++ samples/ZeroAlloc.Cache.AotSmoke/ICustomerService.cs
@@ -9,14 +9,3 @@ public interface ICustomerService
 {
     ValueTask<string> GetNameAsync(int customerId, CancellationToken ct);
 }
-
-public sealed class CustomerService : ICustomerService
-{
-    public int CallCount { get; private set; }
-
-    public ValueTask<string> GetNameAsync(int customerId, CancellationToken ct)
-    {
-        CallCount++;
-        return ValueTask.FromResult($"customer-{customerId}");
-    }
-}
